Update Substrate to master (#349)
* Update to tag Polkadot v0.9.10
* Runtime compiles
* Update Substrate to latest
* Save lock file
* Fix Cargo version
* Fix cargo test
* Authors in Cargo.toml and warnings
* Update CI action
* Enumerate pallets
* Lightsync and change node name
* Update template service file
* Added cargo.lock with frozen substrate
Frozen substrate to 88c64e06471cc12aa9b25290f24d5566bcb5dd82

+++ runtime/src/constants.rs
@@ -21,11 +21,11 @@
 pub mod currency {
     use polkadex_primitives::Balance;
 
-	pub const PDEX: Balance = 100_000_000_000;
+    pub const PDEX: Balance = 100_000_000_000;
     pub const UNITS: Balance = PDEX;
-	pub const DOLLARS: Balance = PDEX; // 100_000_000_000
-	pub const CENTS: Balance = DOLLARS / 100; // 1000_000_000
-	pub const MILLICENTS: Balance = CENTS / 1_000; // 1000_000
+    pub const DOLLARS: Balance = PDEX; // 100_000_000_000
+    pub const CENTS: Balance = DOLLARS / 100; // 1000_000_000
+    pub const MILLICENTS: Balance = CENTS / 1_000; // 1000_000
 
     pub const fn deposit(items: u32, bytes: u32) -> Balance {
         items as Balance * 15 * CENTS + (bytes as Balance) * 6 * CENTS
@@ -67,10 +67,10 @@ pub mod time {
     // NOTE: Currently it is not possible to change the epoch duration after the chain has started.
     //       Attempting to do so will brick block production.
     pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 10 * MINUTES;
-    pub const EPOCH_DURATION_IN_SLOTS: u64 = {
+    pub const EPOCH_DURATION_IN_SLOTS: u32 = {
         const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
 
-        (EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
+        (EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u32
     };
 
     // These time units are defined in number of blocks.
